backend/requirements.txt
webrtcvad==2.0.10
aniso8601==9.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
filelock==3.6.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
httpx==0.23.3
huggingface-hub==0.5.1
HyperPyYAML==1.0.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.22.3
packaging==21.3
pycparser==2.21
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
speechbrain==0.5.11
threadpoolctl==3.1.0
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
gunicorn==20.1.0
torch==1.11.0
torchaudio==0.11.0
transformers==4.19.0
pandas==1.4.3
openai-whisper==20231117
setuptools-rust==1.9.0
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl#sha256=d88c737eb7eb766f730f6a2dcb99dfcdb81623e1e0d89a9c638a2182ac19c52e
spacy==3.7.4
